% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_selection_regression.R
\name{backward_selection_regression}
\alias{backward_selection_regression}
\title{Function to backward select variables and do regression based on selected variables.}
\usage{
backward_selection_regression(
  data,
  response,
  expected_number_of_regressor,
  summary = FALSE
)
}
\arguments{
\item{data}{A data frame type data set for regression and variable selection}

\item{response}{A response variable for your regression in character format, eg:"xxx"}

\item{expected_number_of_regressor}{Expected number of regressors}

\item{summary}{whether do a summary for the regression model based on the selected variables}
}
\value{
The selected variables, summary of such regression model if you set the parameter summary TRUE
}
\description{
This function takes a data set, a response variable and expected number of regressors as input
and returns the selected variable based by backward selection and AIC.
Also returns the summary of corresponding regression model if you set parameter summary TRUE.
}
\examples{
# The following sample data set includes socio-economic information with suicide rates by country in 2011.
data("SuicideData")
backward_selection_regression(SuicideData, "suicides.100k.pop", 3)
backward_selection_regression(SuicideData, "suicides.100k.pop", 4)
backward_selection_regression(SuicideData, "suicides.100k.pop", 3, summary = TRUE)

}
