% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sample_size.R
\name{calculate_sample_size}
\alias{calculate_sample_size}
\title{Calculate Required Sample Size for Estimating a Proportion}
\usage{
calculate_sample_size(p_hat, desired_error, confidence_level = 0.95)
}
\arguments{
\item{p_hat}{Estimated proportion based on initial data or prior knowledge.}

\item{desired_error}{Maximum margin of error for the proportion estimate.}

\item{confidence_level}{Confidence level for the interval estimate, default is 0.95.}
}
\value{
The calculated sample size necessary to achieve the specified error with the given confidence level.
}
\description{
This function estimates the required sample size to achieve a desired error in estimating
a population proportion with a specified confidence level. The function assumes a simple
random sample from a large population.
}
\examples{
required_n = calculate_sample_size(p_hat = 0.5, desired_error = 0.05)
print(required_n)
}
