% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\name{check_distribution}
\alias{check_distribution}
\title{Function to check distribution}
\usage{
check_distribution(data, distribution, detail = FALSE, size = 1000, prob = 0.5)
}
\arguments{
\item{data}{data frame}

\item{distribution}{The distribution you expected}

\item{detail}{Set TRUE to show the details of the test}

\item{size}{Number of trials}

\item{prob}{The probability for the binomial distribution}
}
\value{
Whether the data follows expected distribution
}
\description{
This function takes a data and an expected distribution as input and returns
whether the data follows expected distribution.
}
\examples{
data("SuicideData")
check_distribution(SuicideData$suicides.100k.pop, "normal")
check_distribution(SuicideData$suicides.100k.pop, "normal", detail = TRUE)
set.seed(1)
check_distribution(runif(1000),"uniform", detail = TRUE)
check_distribution(rpois(1000,2), "poisson", detail =TRUE)
check_distribution(rbinom(1000,100,0.5), "binomial", detail =TRUE, size = 100, prob = 0.5)

}
