% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_themer.R
\name{create_xaringan_theme}
\alias{create_xaringan_theme}
\title{Update Xaringan CSS Theme}
\usage{
create_xaringan_theme(
  output = here::here("assets", "css", "xaringan-themer.css")
)
}
\arguments{
\item{output}{the CSS filename (please use \code{here::here()})}
}
\description{
This function updates the Xaringan CSS Theme of a particular talk.
}
\examples{
\dontrun{
create_xaringan_theme()
create_xaringan_theme(
  here::here("courses","docker", "assets", "css", "xaringan-themer.css")
)
}
}
