% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{data_cleaning}
\alias{data_cleaning}
\title{Data Cleaning}
\usage{
data_cleaning(
  data,
  remove_missing = FALSE,
  remove_repeating = FALSE,
  fill_NA = FALSE,
  get_dummies = FALSE
)
}
\arguments{
\item{data}{the data that need cleaning.}

\item{remove_missing}{whether to remove the missing data in the data set.}

\item{remove_repeating}{whether to remove the repeated data in the data set.}

\item{fill_NA}{whether to fill the missing values.}

\item{get_dummies}{whether to create dummy variables.}
}
\value{
the data after the cleaning procedure.
}
\description{
Conduct data cleaning; remove or modify bad data.
}
\examples{
data("SuicideData")
data_cleaning(SuicideData)

}
