% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prefecture.R
\name{get_prefecture}
\alias{get_prefecture}
\title{Get Occitanie Prefectures Coordinates and Population}
\usage{
get_prefecture(region = "Occitanie")
}
\arguments{
\item{region}{just tested with "Occitanie".}
}
\value{
a 5-columns data frame with Department, Prefecture, Population,
Longitude, and Latitude.
}
\description{
This function retrieves coordinates and population of prefectures of the
French region Occitanie using the API Nominatim.
}
\examples{
\dontrun{
get_prefecture()
}
}
