% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_slides.R
\name{knit_slides}
\alias{knit_slides}
\title{Knit Slides}
\usage{
knit_slides(label, home = FALSE, course = TRUE, pdf = FALSE)
}
\arguments{
\item{label}{the course/exercise subfolder name (e.g. "intro-git").}

\item{home}{a boolean. If TRUE, only the Homepage will be re-knit.}

\item{course}{a boolean. If TRUE, searchs in the "courses" folder otherwise
searchs in the "exercises" folder. Ignore if \code{home = TRUE}.}

\item{pdf}{a boolean. If TRUE, converts HTML to PDF.}
}
\description{
This function re-knits Rmd slides to HTML and PDF outputs (stored in the
same folder). You need to have installed the Google Chrome web browser to
create the PDF version. All the Rmd files need to be named \strong{index.Rmd}.
}
\examples{
\dontrun{
# Knit the Homepage ----
knit_slides(home = TRUE)

# Knit Git course ----
knit_slides(label = "intro-git")
knit_slides(label = "intro-git", pdf = FALSE)

# Knit Git exercise ----
knit_slides(label = "intro-git", course = FALSE)
knit_slides(label = "intro-git", course = FALSE, pdf = FALSE)
}
}
