% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outliers_handling.R
\name{Outliers_handling}
\alias{Outliers_handling}
\title{A function to handle outliers in the data set.}
\usage{
Outliers_handling(
  data,
  threshold = 0.5,
  target_column,
  plot_before = TRUE,
  plot_after = TRUE
)
}
\arguments{
\item{data}{the data frame to be processed.}

\item{threshold}{the threshold to determine whether a value is considered as an outlier.}

\item{target_column}{the target column as response.}

\item{plot_before}{whether to plot the original data.}

\item{plot_after}{whether to plot the processed data.}
}
\value{
the processed data and the plots if needed.
}
\description{
This function can process the outliers in a data set.
}
\examples{
data("SuicideData")

}
