% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_analysis.R
\name{plot_data_analysis}
\alias{plot_data_analysis}
\title{Plot Data Analysis}
\usage{
plot_data_analysis(
  data,
  hist = TRUE,
  pair = TRUE,
  hist_col = "lightblue",
  pair_col = "darkgreen"
)
}
\arguments{
\item{data}{A data frame containing numeric variables to be analyzed.}

\item{hist}{Logical. If \code{TRUE}, histograms of numeric variables will be plotted. Default is \code{TRUE}.}

\item{pair}{Logical. If \code{TRUE} and more than one numeric variable exists, pairwise relationships will be plotted. Default is \code{TRUE}.}

\item{hist_col}{Color specification for histograms, default to be lightblue. If provided, the same color will be used for all histograms.}

\item{pair_col}{Color specification for pairwise plots, default to be darkgreen. If provided, the same color will be used for all pairwise plots.}
}
\description{
Plots histograms and pairwise relationships for numeric variables in a data frame.
}
\examples{
data <- data.frame(
  Age = c(25, 30, 35, 40, 45),
  Height = c(175, 180, 165, 170, 160),
  Weight = c(70, 80, 75, 65, 60),
  Gender = as.factor(c("Male", "Female", "Male", "Female", "Male")),
  Income = c(50000, 60000, 55000, 58000, 62000)
)
plot_data_analysis(data)


}
