% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_slides.R
\name{knit_slides}
\alias{knit_slides}
\title{Knit RMarkdown Slides}
\usage{
knit_slides(name)
}
\arguments{
\item{name}{a character of length 1. The course sub-folder name
(e.g. \code{"template"}).}
}
\description{
This function knits Rmd slides to HTML output (stored in the same folder).
All Rmd files need to be named \strong{index.Rmd}.
}
\examples{
\dontrun{
## Knit course ----
knit_slides(name = "template")
}
}
